[ #1346 refactor ] move Fixity type to Syntax.Common
Also: define some more Null instances
