+++ src/full/Agda/Compiler/JS/Compiler.hs
@@ -2,8 +2,10 @@
 module Agda.Compiler.JS.Compiler where
 
 import Prelude hiding ( null, writeFile )
+
 import Control.Monad.Reader ( liftIO )
 import Control.Monad.Trans
+
 import Data.Char ( isSpace )
 import Data.List ( intercalate, partition )
 import Data.Set ( Set, null, insert, difference, delete )
@@ -11,11 +13,16 @@ import Data.Traversable (traverse)
 import Data.Map ( fromList )
 import qualified Data.Set as Set
 import qualified Data.Map as Map
+
 import System.Directory ( createDirectoryIfMissing )
 import System.FilePath ( splitFileName, (</>) )
 
+import Paths_Agda
+
+import Agda.Interaction.Options
 import Agda.Interaction.Imports ( isNewerThan )
-import Agda.Syntax.Common ( Nat, unArg, NameId(..) )
+
+import Agda.Syntax.Common
 import Agda.Syntax.Concrete.Name ( isNoName )
 import Agda.Syntax.Abstract.Name
   ( ModuleName, QName,
@@ -26,11 +33,13 @@ import Agda.Syntax.Internal
 import Agda.Syntax.Literal ( Literal(..) )
 import Agda.Syntax.Fixity
 import qualified Agda.Syntax.Treeless as T
+
 import Agda.TypeChecking.Monad
 import Agda.TypeChecking.Monad.Builtin
 import Agda.TypeChecking.Monad.Debug ( reportSLn )
 import Agda.TypeChecking.Reduce ( instantiateFull )
 import Agda.TypeChecking.Pretty
+
 import Agda.Utils.Maybe
 import Agda.Utils.Monad ( (<$>), (<*>), ifM )
 import Agda.Utils.Pretty (prettyShow)
@@ -52,10 +61,6 @@ import Agda.Compiler.JS.Syntax
 import Agda.Compiler.JS.Substitution
   ( curriedLambda, curriedApply, emp, apply )
 import qualified Agda.Compiler.JS.Pretty as JSPretty
-
-import Agda.Interaction.Options
-
-import Paths_Agda
 
 import Agda.Utils.Impossible (__IMPOSSIBLE__)
