+++ src/full/Agda/Syntax/Concrete/Fixity.hs
@@ -17,11 +17,12 @@ import qualified Data.Set as Set
 import Data.Semigroup
 #endif
 
-import Agda.Syntax.Position
+import Agda.Syntax.Builtin (builtinsNoDef)
+import Agda.Syntax.Common
+import Agda.Syntax.Concrete
 import Agda.Syntax.Fixity
 import Agda.Syntax.Notation
-import Agda.Syntax.Concrete
-import Agda.Syntax.Builtin (builtinsNoDef)
+import Agda.Syntax.Position
 import Agda.TypeChecking.Positivity.Occurrence (Occurrence)
 
 import Agda.Utils.Functor
@@ -52,11 +53,11 @@ plusFixities m1 m2
   where
     --  Merge two fixities, assuming there is no conflict
     mergeFixites name (Fixity' f1 s1 r1) (Fixity' f2 s2 r2) = Fixity' f s $ fuseRange r1 r2
-              where f | f1 == noFixity = f2
-                      | f2 == noFixity = f1
+              where f | null f1 = f2
+                      | null f2 = f1
                       | otherwise = __IMPOSSIBLE__
-                    s | s1 == noNotation = s2
-                      | s2 == noNotation = s1
+                    s | null s1 = s2
+                      | null s2 = s1
                       | otherwise = __IMPOSSIBLE__
 
     -- Compute a list of conflicts in a format suitable for error reporting.
@@ -65,8 +66,8 @@ plusFixities m1 m2
 
     -- Check for no conflict.
     compatible (Fixity' f1 s1 _) (Fixity' f2 s2 _) =
-      (f1 == noFixity   || f2 == noFixity  ) &&
-      (s1 == noNotation || s2 == noNotation)
+      (null f1 || null f2) &&
+      (null s1 || null s2)
 
 -- | While 'Fixities' and Polarities are not semigroups under disjoint
 --   union (which might fail), we get a semigroup instance for the
@@ -238,4 +239,3 @@ declaredNames d = case d of
   Pragma (BuiltinPragma _ b (QName x))
     | b `elem` builtinsNoDef -> declaresName x
   Pragma{}             -> mempty
-
